// src/modules/core/api/types/Farming.ts
export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export type ClaimedReward = ContractEventLog<{
  caller: string;
  claimer: string;
  receiver: string;
  amount: string;
  0: string;
  1: string;
  2: string;
  3: string;
}>;
export type Deposit = ContractEventLog<{
  user: string;
  pid: string;
  amount: string;
  0: string;
  1: string;
  2: string;
}>;
export type EmergencyWithdraw = ContractEventLog<{
  user: string;
  pid: string;
  amount: string;
  0: string;
  1: string;
  2: string;
}>;
export type Initialized = ContractEventLog<{
  version: string;
  0: string;
}>;
export type OwnershipTransferred = ContractEventLog<{
  previousOwner: string;
  newOwner: string;
  0: string;
  1: string;
}>;
export type Withdraw = ContractEventLog<{
  user: string;
  pid: string;
  amount: string;
  0: string;
  1: string;
  2: string;
}>;

export interface Farming extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): Farming;
  clone(): Farming;
  methods: {
    addPool(
      _token: string,
      _strategy: string,
      _withUpdate: boolean
    ): NonPayableTransactionObject<string>;

    blockThirdPartyActions(arg0: string): NonPayableTransactionObject<boolean>;

    claim(
      _user: string,
      _pids: (number | string | BN)[]
    ): NonPayableTransactionObject<string>;

    claimReceiver(arg0: string): NonPayableTransactionObject<string>;

    claimableReward(
      _user: string,
      _pids: (number | string | BN)[]
    ): NonPayableTransactionObject<string[]>;

    deposit(
      _pid: number | string | BN,
      _wantAmt: number | string | BN,
      _claimRewards: boolean,
      _userAddress: string
    ): NonPayableTransactionObject<string>;

    emergencyWithdraw(
      _pid: number | string | BN
    ): NonPayableTransactionObject<void>;

    inCaseTokensGetStuck(
      _token: string,
      _amount: number | string | BN
    ): NonPayableTransactionObject<void>;

    incentiveVoting(): NonPayableTransactionObject<string>;

    initialize(
      _rewardToken: string,
      _incentiveVoting: string
    ): NonPayableTransactionObject<void>;

    massUpdatePools(): NonPayableTransactionObject<void>;

    owner(): NonPayableTransactionObject<string>;

    poolInfo(arg0: number | string | BN): NonPayableTransactionObject<{
      token: string;
      strategy: string;
      rewardsPerSecond: string;
      lastRewardTime: string;
      accRewardPerShare: string;
      0: string;
      1: string;
      2: string;
      3: string;
      4: string;
    }>;

    poolLength(): NonPayableTransactionObject<string>;

    renounceOwnership(): NonPayableTransactionObject<void>;

    rewardToken(): NonPayableTransactionObject<string>;

    setBlockThirdPartyActions(
      _block: boolean
    ): NonPayableTransactionObject<void>;

    setClaimReceiver(_receiver: string): NonPayableTransactionObject<void>;

    stakedWantTokens(
      _pid: number | string | BN,
      _user: string
    ): NonPayableTransactionObject<string>;

    startTime(): NonPayableTransactionObject<string>;

    transferOwnership(newOwner: string): NonPayableTransactionObject<void>;

    updatePool(_pid: number | string | BN): NonPayableTransactionObject<string>;

    userInfo(
      arg0: number | string | BN,
      arg1: string
    ): NonPayableTransactionObject<{
      shares: string;
      rewardDebt: string;
      claimable: string;
      0: string;
      1: string;
      2: string;
    }>;

    withdraw(
      _pid: number | string | BN,
      _wantAmt: number | string | BN,
      _claimRewards: boolean
    ): NonPayableTransactionObject<string>;

    withdrawAll(
      _pid: number | string | BN,
      _claimRewards: boolean
    ): NonPayableTransactionObject<string>;
  };
  events: {
    ClaimedReward(cb?: Callback<ClaimedReward>): EventEmitter;
    ClaimedReward(
      options?: EventOptions,
      cb?: Callback<ClaimedReward>
    ): EventEmitter;

    Deposit(cb?: Callback<Deposit>): EventEmitter;
    Deposit(options?: EventOptions, cb?: Callback<Deposit>): EventEmitter;

    EmergencyWithdraw(cb?: Callback<EmergencyWithdraw>): EventEmitter;
    EmergencyWithdraw(
      options?: EventOptions,
      cb?: Callback<EmergencyWithdraw>
    ): EventEmitter;

    Initialized(cb?: Callback<Initialized>): EventEmitter;
    Initialized(
      options?: EventOptions,
      cb?: Callback<Initialized>
    ): EventEmitter;

    OwnershipTransferred(cb?: Callback<OwnershipTransferred>): EventEmitter;
    OwnershipTransferred(
      options?: EventOptions,
      cb?: Callback<OwnershipTransferred>
    ): EventEmitter;

    Withdraw(cb?: Callback<Withdraw>): EventEmitter;
    Withdraw(options?: EventOptions, cb?: Callback<Withdraw>): EventEmitter;

    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };

  once(event: "ClaimedReward", cb: Callback<ClaimedReward>): void;
  once(
    event: "ClaimedReward",
    options: EventOptions,
    cb: Callback<ClaimedReward>
  ): void;

  once(event: "Deposit", cb: Callback<Deposit>): void;
  once(event: "Deposit", options: EventOptions, cb: Callback<Deposit>): void;

  once(event: "EmergencyWithdraw", cb: Callback<EmergencyWithdraw>): void;
  once(
    event: "EmergencyWithdraw",
    options: EventOptions,
    cb: Callback<EmergencyWithdraw>
  ): void;

  once(event: "Initialized", cb: Callback<Initialized>): void;
  once(
    event: "Initialized",
    options: EventOptions,
    cb: Callback<Initialized>
  ): void;

  once(event: "OwnershipTransferred", cb: Callback<OwnershipTransferred>): void;
  once(
    event: "OwnershipTransferred",
    options: EventOptions,
    cb: Callback<OwnershipTransferred>
  ): void;

  once(event: "Withdraw", cb: Callback<Withdraw>): void;
  once(event: "Withdraw", options: EventOptions, cb: Callback<Withdraw>): void;
}